requirements.txt
altgraph==0.10.2
arrow==0.7.0
backports-abc==0.4
backports.ssl-match-hostname==3.4.0.2
bdist-mpkg==0.5.0
BeautifulSoup==3.2.1
beautifulsoup4==4.4.1
bonjour-py==0.3
bson==0.4.1
certifi==2015.9.6.2
coverage==4.0.1
disqus==0.0.4
Django==1.8.7
django-debug-toolbar==1.4
django-disqus==0.5
django-emoticons==1.1.2
django-forms-bootstrap==3.0.1
django-paypal==0.2.7
django-tinymce==2.2.0
Flask==0.10.1
funcsigs==0.4
hacker-news==2.1.4
hackernews-scraper==1.0.2
itsdangerous==0.24
Jinja2==2.8
lxml==3.5.0
macholib==1.5.1
MarkupSafe==0.23
matplotlib==1.3.1
mock==1.3.0
modulegraph==0.10.4
MySQL-python==1.2.5
nose==1.3.7
numpy==1.8.0rc1
oauthlib==1.0.3
pandas==0.17.0
pandas-datareader==0.2.0
pbr==1.8.1
Pillow==2.9.0
prettytable==0.7.2
py2app==0.7.3
pymongo==3.1
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==0.13.1
pyparsing==2.0.1
python-dateutil==1.5
pytz==2013.7
requests==2.8.1
requests-oauthlib==0.5.0
scipy==0.13.0b1
Scrapy==0.16.4
simplejson==3.8.0
singledispatch==3.4.0.3
six==1.10.0
sqlparse==0.1.18
stripe==1.27.1
tornado==4.3
tweepy==3.4.0
Twisted==15.4.0
vboxapi==1.0
virtualenv==13.1.2
w3lib==1.13.0
Werkzeug==0.11.2
wheel==0.26.0
xattr==0.6.4
zope.interface==4.1.1
